fix: preco response decimal
- Preço na response como decimal para não arrendondar valores tipo "249,99"
- Range de preço no Create e Update atualizados para melhor refletir a lógica do limite do próprio decimal e do banco de dados

+++ Application/Responses/Produto/ProdutoResponse.cs
@@ -6,7 +6,7 @@ namespace Application.Responses.Produto
         int Id,
         string Nome,
         string Descricao,
-        int Preco,
+        decimal Preco,
         CategoriaResponse Categoria
     );
 }
